automation/services/mina-bp-stats/leaderboard-bot_v2/requirements.txt
kubernetes
python-dotenv
slackclient
psycopg2